refactor(sql): 修改支付订单表字段类型
- 将支付订单表中的 id 字段类型从 serial 修改为 bigserial，以支持更大的 ID 值
- 将 user_id 字段类型从 integer 修改为 bigint，以提高用户 ID 的存储能力
- 保持其他字段及表结构不变，确保数据兼容性

+++ sql/ddl.sql
@@ -93,8 +93,8 @@ create type order_status as enum ('created', 'paid', 'closed');
 
 -- 支付订单表
 create table if not exists pay_order (
-    id serial primary key,
-    user_id integer not null,
+    id bigserial primary key,
+    user_id bigint not null,
     level order_level not null,
     edition product_edition not null,
     pay_from pay_from not null,
